Moved IndustryGrid to the Industries section composition
