+++ src/compositions/pages/Home/Industries/IndustryGrid/index.jsx
