+++ src/compositions/pages/Home/Industries/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import IndustryGrid from '../../../../components/IndustryGrid';
+import IndustryGrid from './IndustryGrid';
 import PageSection from '../../../../components/PageSection';
 import PageSectionHeader from '../../../../components/PageSectionHeader';
